Changes to support Lua 5.2 (previously only supported 5.1). Tables are now stored to the registry table instead of the global table by default.
Lua 5.2 removed LUA_GLOBALSINDEX which prevented the code from compiling/working. The easiest solution s just to use LUA_REGISTRYINDEX. For the primary design case, this is fine. But in the hypothetical (but interesting) case where you might want to try to access everything you have set in LuaHashMap directly in Lua script, the global table is an interesting option to have.

So I refactored the code so you can toggle how you compile the library in case you want to activate this feature. For Lua 5.2, the code to get to the global table is a little more cumbersome, so theoretically there might be a very slight performance loss due to the few extra lines of code I have to issue for the global table case.

For the registry table, the code for 5.1 and 5.2 is exactly the same and was a simple change of using LUA_REGISTRYINDEX instead of LUA_GLOBALSINDEX.

Use the #define LUA_HASHMAP_USE_GLOBAL_TABLE when building to use the global table.