Fix to handle case where removing while iterating breaks because of garbage collection. Note that the contract for Removing while iterating is as follows:
Removing in an iterator loop is only safe with iterator functions. Use LuaHashMap_RemoveAtIterator and LuaHashMap_IteratorNext. Don't remove using other remove functions otherwise the iterator could be left using a dangling reference.

After removing, the values in the iterator are invalid. The only function you should call on it is IteratorNext which has special handling to find the next position.



RemoveAtIterator will now lookup and cache the next key and save it in the iterator. IteratorNext will detect this value and recover it by using a new flag called isNext.

There is a corner case where the cached values could be stale when the user is expecting them to be fresh.
	    hash_iterator = LuaHashMap_GetIteratorAtBegin(hash_map);
	    do
	    {
	        fprintf(stderr, "Removing: Price of %s: %lf\n",
	            LuaHashMap_GetCachedKeyStringAtIterator(&hash_iterator),
	            LuaHashMap_GetCachedValueNumberAtIterator(&hash_iterator));
	        LuaHashMap_RemoveAtIterator(&hash_iterator);
	        // Change a value in the hash that happens to be the next position.
	        LuaHashMap_SetValue*ForKey*(...);

	    } while(LuaHashMap_IteratorNext(&hash_iterator));

In this case, the value was already cached in RemoveAtIterator so in the next loop, the value is stale.
However, since this technically wasn't using the iterator APIs to set the value, this is one of the cases GetCached is documented to miss. For performance reasons, I don't really want to re-fetch the value here. I presume the common case is to iterate and remove so it doesn't make sense to do a double lookup.



Changed the union type so key and values could share the same one. Reorganized the struct layout to hopefully be less fragmented.

Added some test cases for NULL checking and this iterator problem.


Probably should add checks for isNext in all the public APIs next to abort out if true.