Optimizations: Added new auxiliary APIs for passing explicit string lengths. Note: There is an API behavior change. Passing in a NULL string for the value no longer removes the entry from the hash. Instead it sets an empty string. This is a result of passing explicit string lengths as described below. This is probably better anyway as it is more consistent with setting Pointers to NULL.
My theory is that the strlen calls internal to lua_pushstring may be responsible for much of the lost time in my extreme benchmarks. Since Lua has special/additional APIs to explicitly pass string length, I've exposed them as alternative APIs that end with the 'WithLength' suffix.

In addition, the LuaHashMapIterator now contains a StringContainer struct which holds both the string pointer and string length so iterator access should remain fast. The StringContainer is now theString part of the union in the iterator. This will theoretically save memory when the size of lua_Number is greater than or equal to the size of the string pointer+size_t. This is generally true in 32-bit land (8==4+4), though not so much for 64-bit unless long double is defined as the number type.

Finally, I changed my internal safestrcmp function to a safestrequal. I realized I don't need the greater-than/less-than comparisons that strcmp offers and I just need a boolean. So now that I have lengths and don't need inequality comparisons, I can do faster string compares.

I need to write a benchmark that exploits the new string length APIs. However, I think the changes have already improved the benchmark scores, particularly with the iterators and the new strequal function.