Discovered the _Generic types are very strict so things like char* and const char* are different. So I added a whole bunch more rules and filled in all the integer and floating point types for 'value'.
In addition, there is a problem with string literals where they are really const char[] but there doesn't seem to be a way to express this because the type is incomplete. (I tried an explicit number in the array, and while it compiled, the number didn't actually seem to let it resolve to the correct type even when the lengths matched.) Maybe this is a clang bug.

More info is found here:
http://www.robertgamble.net/2012/01/c11-generic-selections.html