I couldn't leave well-enough alone. Did one more pass for restrict and const to add them where I think appropriate.
restrict was added in a few more places, most notable size_t* (return length through an argument).

I did intentionally skip the void* (pointer) APIs because I think in theory, you could be adding the LuaHashMap pointer to the table itself, thus the pointers are the same/aliased.

I added const to the GetKey and GetCachedValue functions (which are already inlined). While I doubt the compiler will generate any better code (and the inline part should already be more than sufficient), this is now a contract for users of the API in that I'm not changing the iterator struct. This implies that a struct copy before one of these const operations is the same struct as after.

LuaHashMapIterator the_iterator = LuaHashMap_GetIterator(?);
LuaHashMapIterator saved_iterator = the_iterator;
LuaHashMapIterator_GetCachedValue(&the_iterator);
// The const contract now implies the_iterator is still the same as saved_iterator.